Make bit-search autofocus
