+++ libs/components/src/input/index.ts
@@ -1,1 +1,2 @@
 export * from "./input.module";
+export * from "./autofocus.directive";
