+++ libs/components/src/search/search.component.html
@@ -8,6 +8,7 @@
     <i class="bwi bwi-search bwi-fw tw-text-muted"></i>
   </label>
   <input
+    #input
     bitInput
     type="search"
     [id]="id"
