+++ libs/components/src/search/search.component.ts
@@ -1,5 +1,7 @@
-import { Component, Input } from "@angular/core";
+import { Component, ElementRef, Input, ViewChild } from "@angular/core";
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from "@angular/forms";
+
+import { FocusableElement } from "../input/autofocus.directive";
 
 let nextId = 0;
 
@@ -12,17 +14,27 @@ let nextId = 0;
       multi: true,
       useExisting: SearchComponent,
     },
+    {
+      provide: FocusableElement,
+      useExisting: SearchComponent,
+    },
   ],
 })
-export class SearchComponent implements ControlValueAccessor {
+export class SearchComponent implements ControlValueAccessor, FocusableElement {
   private notifyOnChange: (v: string) => void;
   private notifyOnTouch: () => void;
+
+  @ViewChild("input") private input: ElementRef<HTMLInputElement>;
 
   protected id = `search-id-${nextId++}`;
   protected searchText: string;
 
   @Input() disabled: boolean;
   @Input() placeholder: string;
+
+  focus() {
+    this.input.nativeElement.focus();
+  }
 
   onChange(searchText: string) {
     if (this.notifyOnChange != undefined) {
